BugFix: Previous versions of CyREST created conflicts for some resources
Due to CyREST providing packages and bundles, previous installs of
CyREST could cause conflicting package resolutions. A restart was
required.

BugFix introduces a check to see if a previous version was installed,
and notifies the user via a dialog if a restart is required.

Added some tests for OSGi JAX RS Connector bundle loading, and the start
of testing for the BugFix.