Groups Swagger Update + BugFix + Code Cleanup
Found a bug in getting a specific group: the  fixed it in the process of figuring out what it was. group.getGroupNode().getSUID() == suid was being treated as a Long to Long equals, which would only be resolved if they were the same Object, instead of having the same long value.