Initial Commit - see Description
Upgrade to Jersey 2.27

Note, this works on a fork of the original OSGi-JAX-RS Connector. Fork is here: https://github.com/dotasek/osgi-jax-rs-connector

The above needs to be built in order for this build to work.

Eventually, this needs to be switched to the cytoscape projects so that Cytoscape can take ownership and create an automated build.